worker/requirements.txt
# Actually installed: pyyaml pyproj requests rq numpy scikit-image
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cycler==0.10.0
idna==3.3
imageio==2.9.0
kiwisolver==1.3.2
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.2
Pillow==8.4.0
pyparsing==2.4.7
pyproj==3.2.1
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==6.0
redis==3.5.3
requests==2.26.0
rq==1.10.0
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
tifffile==2021.10.12
urllib3==1.26.7
